+++ sperl_dumper.h
@@ -6,7 +6,7 @@
 void SPerl_DUMPER_dump_sperl(SPerl* sperl);
 void SPerl_DUMPER_dump_classs(SPerl* sperl, SPerl_ARRAY* classs);
 void SPerl_DUMPER_dump_constants(SPerl* sperl, SPerl_ARRAY* constants);
-void SPerl_DUMPER_dump_const_pool(SPerl* sperl, int32_t* const_pool, int32_t size);
+void SPerl_DUMPER_dump_const_pool(SPerl* sperl, int64_t* const_pool, int64_t size);
 void SPerl_DUMPER_dump_constant(SPerl* sperl, SPerl_CONSTANT* constant);
 void SPerl_DUMPER_dump_field(SPerl* sperl, SPerl_FIELD* field);
 void SPerl_DUMPER_dump_sub(SPerl* sperl, SPerl_SUB* sub);
@@ -16,6 +16,6 @@ void SPerl_DUMPER_dump_packages(SPerl* sperl, SPerl_ARRAY* packages);
 void SPerl_DUMPER_dump_resolved_types(SPerl* sperl, SPerl_ARRAY* resolved_types);
 void SPerl_DUMPER_dump_ast(SPerl* sperl, SPerl_OP* op);
 void SPerl_DUMPER_dump_constant_pool(SPerl* sperl, SPerl_CONSTANT_POOL* constant_pool);
-void SPerl_DUMPER_dump_bytecode_array(SPerl* sperl, SPerl_BYTECODE_ARRAY* bytecode_array, int32_t base, int32_t length);
+void SPerl_DUMPER_dump_bytecode_array(SPerl* sperl, SPerl_BYTECODE_ARRAY* bytecode_array, int64_t base, int64_t length);
 
 #endif
